- Atributos nuevos para entidades. - Dtos modificados con estos nuevos atributos para el apartado response. - Implementación correcta de las reglas del servicio 6,7,8 y 9. - testeo de la logica detrás de estos servicios. - correción en pequeños detalles de los mapeos para cumplir con la logica nueva dentro de los servicios.

+++ proyecto_final/src/main/java/com/unimag/edu/proyecto_final/api/dto/BaggageDtos.java
@@ -1,9 +1,10 @@
 package com.unimag.edu.proyecto_final.api.dto;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 public class BaggageDtos {
-    public record BaggageCreateRequest(Long ticketId,Double weightKg, Double fee, String tagCode) implements Serializable {}
-    public record BaggageResponse(Long id, Long ticketId, Double weightKg, Double fee, String tagCode) implements Serializable {}
-    public record BaggageUpdateRequest(Double weight, Double fee) implements Serializable {}
+    public record BaggageCreateRequest(Long ticketId,Double weightKg, BigDecimal fee, String tagCode) implements Serializable {}
+    public record BaggageResponse(Long id, Long ticketId, Double weightKg, BigDecimal fee, String tagCode) implements Serializable {}
+    public record BaggageUpdateRequest(Double weight, BigDecimal fee) implements Serializable {}
 }
